COMPMID-2486: Remove disabled compiler warnings
Removed the following flags:
-Wno-format-nonliteral: This had a side effect on
Error.h that resulted in rewriting most of the macros. Since
I was at it I removed all the va_args in order to comply
with DCL50-CPP.
-Wno-deprecated-increment-bool
-Wno-vla-extension
-Wno-mismatched-tags
-Wno-redundant-move

Reviewed-on: https://review.mlplatform.org/c/2069
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>

+++ arm_compute/core/ITensorInfo.h
@@ -148,7 +148,7 @@ public:
      *
      * @return Offset in bytes from the beginning of the memory allocation to access the element (x, y, z, ...)
      */
-    virtual size_t offset_element_in_bytes(const Coordinates &pos) const = 0;
+    virtual int32_t offset_element_in_bytes(const Coordinates &pos) const = 0;
 
     /** Element size in bytes calculated as data_size() * num_channels()
      *
